Rework API system (Add Route)
