+++ .env
@@ -2,25 +2,26 @@ VITE_API_CONTACT = http://localhost:3001/contact
 VITE_API_MAIN = http://localhost:3001/main
 VITE_API_INFO = http://localhost:3001/info
 
-VITE_API_CAR_ADD = http://localhost:3001/CarAdd
-VITE_API_CAR_EDIT = http://localhost:3001/CarEdit
-VITE_API_CAR_DELETE = http://localhost:3001/CarDelete
-VITE_API_CAR_MAIN = http://localhost:3001/CarMain
-VITE_API_CAR_DETAIL = http://localhost:3001/CarDetail
+VITE_API_CAR_ADD = http://localhost:3001/Car/CarAdd
+VITE_API_CAR_EDIT = http://localhost:3001/Car/CarEdit
+VITE_API_CAR_DELETE = http://localhost:3001/Car/CarDelete
+VITE_API_CAR_MAIN = http://localhost:3001/Car/CarMain
+VITE_API_CAR_DETAIL = http://localhost:3001/Car/CarDetail
 
 
-VITE_API_CUSTOMER_ADD = http://localhost:3001/CustomerAdd
-VITE_API_CUSTOMER_EDIT = http://localhost:3001/CustomerEdit
-VITE_API_CUSTOMER_MAIN = http://localhost:3001/CustomerMain
+VITE_API_CUSTOMER_ADD = http://localhost:3001/Customer/CustomerAdd
+VITE_API_CUSTOMER_EDIT = http://localhost:3001/Customer/CustomerEdit
+VITE_API_CUSTOMER_MAIN = http://localhost:3001/Customer/CustomerMain
 
-VITE_API_RENT_MAIN = http://localhost:3001/RentMain
-VITE_API_RENT_ADD = http://localhost:3001/RentAdd
-VITE_API_RENT_EDIT = http://localhost:3001/RentEdit
-VITE_API_RENT_CHECKOUT = http://localhost:3001/RentCheckout
+VITE_API_RENT_MAIN = http://localhost:3001/Rent/RentMain
+VITE_API_RENT_ADD = http://localhost:3001/Rent/RentAdd
+VITE_API_RENT_EDIT = http://localhost:3001/Rent/RentEdit
+VITE_API_RENT_CHECKOUT = http://localhost:3001/Rent/RentCheckout
 
-VITE_API_RENT_DETAIL = http://localhost:3001/RentDetail
+VITE_API_RENT_DETAIL = http://localhost:3001/Rent/RentDetail
 
-VITE_API_BOOK_ADD = http://localhost:3001/BookAdd
-VITE_API_BOOK_CANCEL = http://localhost:3001/BookCancel
-VITE_API_BOOK_DETAIL = http://localhost:3001/BookDetail
-VITE_API_BOOK_CREATE_RENT = http://localhost:3001/BookCreateRent
+VITE_API_BOOK_ADD = http://localhost:3001/Book/BookAdd
+VITE_API_BOOK_EDIT = http://localhost:3001/Book/BookEdit
+VITE_API_BOOK_CANCEL = http://localhost:3001/Book/BookCancel
+VITE_API_BOOK_DETAIL = http://localhost:3001/Book/BookDetail
+VITE_API_BOOK_CREATE_RENT = http://localhost:3001/Book/BookCreateRent
