+++ server/package.json
@@ -16,6 +16,7 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-fileupload": "^1.4.2",
+    "express-router": "^0.0.1",
     "firebase": "^10.5.2",
     "imgur": "^1.0.2",
     "mongodb": "^6.1.0",
