+++ server/server.js
@@ -2,6 +2,11 @@ const check = require("./checkConnection");
 const contactModel = require("./models/Contact");
 const [result, port, app] = check(true);
 
+const XeRoute = require('./routes/Xe')
+const SoDatXeRoute = require('./routes/SoDatXe');
+const SoXeRoute = require('./routes/SoXe')
+const KhachHangRoute = require('./routes/KhachHang');
+
 const XeModel = require("./models/Xe");
 const KhachHangModel = require("./models/KhachHang");
 const SoXeModel = require("./models/SoXe");
@@ -53,210 +58,17 @@ if(result){
     })
     
     //--------- Xử lý quản lý xe ---------///
-    
-    app.post('/CarAdd/', async (req,res) => {
-        await XeModel.create(req.body)
-        .then(() => res.json({success: true, msg: 'Thêm xe thành công'}))
-        .catch(() => res.json({ success: false, msg: 'Thêm xe thất bại. Vui lòng thử lại sau !' }))
-    })
-
-
-    app.post('/CarDelete/:ID', async (req, res) => {
-        await SoXeModel.deleteMany({IDXe : `${req.params.ID}`})
-        await XeModel.deleteOne({ _id : `${req.params.ID}`})
-
-        .then(() => res.json({ success: true, msg: 'Xoá xe thành công' }))
-        .catch(() => res.json({ success: false, msg: 'Xoá xe thất bại. Vui lòng thử lại sau !' }))
-    });
-
-    
-    app.post('/CarEdit/:ID', async (req, res) => {
-        const {TenXe, BienSo, SoCho, TruyenDong, NhienLieu, MoTa, SoTien, HinhAnh, TinhTrang} = req.body;
-        await XeModel.updateOne({ _id : `${req.params.ID}`},{
-            $set: {
-                TenXe : TenXe,
-                BienSo :  BienSo,
-                SoCho : SoCho,
-                TruyenDong : TruyenDong,
-                NhienLieu : NhienLieu,
-                MoTa : MoTa,
-                SoTien : SoTien,
-                HinhAnh : HinhAnh,
-                TinhTrang : TinhTrang,
-            }
-        })
-        .then(() => res.json({ success: true, msg: 'Cập nhật thành công !' }))
-        .catch(() => res.json({ success: false, msg: 'Cập nhật thất bại. Vui lòng thử lại sau !' }))
-    });
-
-    app.get("/CarMain", async (req,res) => {
-        await XeModel.find().then(info => res.json(info)).catch(err => res.json(err))
-    })
-
-    app.get("/CarDetail/:IDXe", async (req,res) => {
-        const info = await XeModel.findOne({_id : `${req.params.IDXe}`}).populate("IDDon")
-        res.json(info)
-    })
+    app.use('/Car', XeRoute)
 
     //--------- Xử lý quản lý sổ đặt xe ---------///
-    app.get('/BookDetail/:IDXe/:IDDon', async (req,res) => {
-        const info = await SoXeModel.find({ 
-            IDXe : `${req.params.IDXe}`,
-            _id : { $nin: [req.params.IDDon] }
-        }).populate("IDXe").populate("IDKH")
-        res.json(info)
-        // .catch(err => res.json(err))
-    })
-
-    app.get('/BookDetail/:IDXe/', async (req,res) => {
-        const info = await SoXeModel.find({ IDXe : `${req.params.IDXe}`,}).populate("IDXe").populate("IDKH")
-        res.json(info)
-        // .catch(err => res.json(err))
-    })
-
-    app.post('/BookAdd/:IDXe/', async (req,res) => {
-        try{
-            await SoXeModel.create(req.body)
-            .then((e) => res.json({success: true, msg: e._id}))
-        }
-        catch{() => {res.json({ success: false, msg: 'Tạo đơn đặt trước xe thất bại. Vui lòng thử lại sau !' })}}
-    })
-
-    app.post('/BookCancel/:IDDon', async (req,res) => {
-        await SoXeModel.updateOne({ _id : `${req.params.IDDon}`},{
-            $set: {TinhTrang : "Huỷ đơn"}
-        })
-        .then(() => res.json({ success: true, msg: 'Huỷ đơn đặt trước thành công !' }))
-        .catch(() => res.json({ success: false, msg: 'Huỷ đơn đặt trước thất bại. Vui lòng thử lại sau !' }))
-    })
-
-    app.post('/BookCreateRent/:IDXe/:IDDon', async (req,res) => {
-        const {TinhTrang, KhachTra} = req.body;
-
-        try{
-            await SoXeModel.updateOne({ _id : `${req.params.IDDon}`},{
-                $set: {
-                  KhachTra : KhachTra,
-                  TinhTrang : TinhTrang
-                }
-            })    
-            .then(async() => {await XeModel.updateOne({ _id : `${req.params.IDXe}`},{
-                    $set: {
-                        TinhTrang : "Đang thuê",
-                        IDDon: new ObjectId(`${req.params.IDDon}`),
-                    }
-                })
-                .then(() => res.json({success: true, msg: "Tạo đơn thành công"}))
-        }) 
-        }
-        catch{() => res.json({ success: false, msg: 'Tạo đơn thất bại. Vui lòng thử lại sau !' })}
-    })
+    app.use('/Book', SoDatXeRoute);
 
     //--------- Xử lý quản lý sổ xe (Khách hàng) ---------//
-    app.post('/CustomerAdd/', async (req,res) => {
-        await KhachHangModel.create(req.body)
-        .then((KhachHangInfo) => res.json({success: true, msg: `${KhachHangInfo._id}`}))
-        .catch(() => res.json({ success: false, msg: 'Thêm khách hàng thất bại. Vui lòng thử lại sau !' }))
-    })
-
-    app.post('/CustomerEdit/:ID', async (req,res) => {
-        const {TenKH, NgaySinh, DiaChi, SoDienThoai, CMND, HinhCMND, BangLai, HinhBangLai} = req.body;
-
-        await KhachHangModel.updateOne({ _id : `${req.params.ID}`},{
-            $set: {
-                TenKH: TenKH, 
-                NgaySinh : NgaySinh, 
-                DiaChi : DiaChi, 
-                SoDienThoai : SoDienThoai, 
-                CMND : CMND, 
-                HinhCMND : HinhCMND, 
-                BangLai : BangLai, 
-                HinhBangLai : HinhBangLai
-            }
-        })
-        .then(() => res.json({ success: true, msg:`${req.params.ID}` }))
-        .catch(() => res.json({ success: false, msg: 'Cập nhật thất bại. Vui lòng thử lại sau !' }))
-
-    })
-
-    app.get('/CustomerMain/', async (req,res) => {
-        await KhachHangModel.find()
-        .then(info => res.json(info))
-        .catch(err => res.json(err))
-    })
-
+    app.use('/Customer', KhachHangRoute);
     
     //--------- Xử lý quản lý sổ xe (Sổ xe) ---------//
-    app.get('/RentMain', async (req,res) => {
-        try{
-            const info = await SoXeModel.find().populate("IDXe").populate("IDKH")
-            res.json(info)
-        }
-        catch(e){res.json(e)}
-    })
-
-    app.post('/RentAdd/:IDXe/', async (req,res) => {
-        req.body.IDXe = new ObjectId(`${req.params.IDXe}`);
-        req.body.IDKH = new ObjectId(`${req.body.IDKH}`);
-        try{
-            await SoXeModel.create(req.body)
-            .then((SoXeInfo) => {
-                    XeModel.updateOne({ _id : `${req.params.IDXe}`},{
-                        $set: {
-                            TinhTrang : "Đang thuê",
-                            IDDon: new ObjectId(`${SoXeInfo._id}`),
-                        }
-                    })
-                    .then(() => res.json({success: true, msg: "Tạo đơn thành công"}))
-            }) 
-        }
-        catch{() => {res.json({ success: false, msg: 'Thêm xe thất bại. Vui lòng thử lại sau !' })}}
-    })
-
-    app.post('/RentEdit/:IDXe/:IDDon', async (req,res) => {
-        const {NgayKetThuc,TinhTrang} = req.body;
-
-        await SoXeModel.updateOne({ _id : `${req.params.IDDon}`},{
-            $set: {
-              NgayKetThuc : NgayKetThuc,
-              TinhTrang : TinhTrang
-            }
-        })
-        .then(() => res.json({ success: true, msg: 'Cập nhật thành công !' }))
-        .catch(() => res.json({ success: false, msg: 'Cập nhật thất bại. Vui lòng thử lại sau !' }))
-    })
-
-    app.post('/RentCheckout/:IDXe/:IDDon', async (req,res) => {
-        
-        await SoXeModel.updateOne({ _id : `${req.params.IDDon}`},{
-            $set: {
-                KhachTra : `${req.body.KhachTra}`,
-                TinhTrang : "Hoàn thành",
-                
-            }
-        })
-
-        await XeModel.updateOne({ _id : `${req.params.IDXe}`},{
-            $set: {TinhTrang : "Còn trống", IDDon : null}
-        })
-        .then(() => res.json({ success: true, msg: 'Trả xe thành công !' }))
-        .catch(() => res.json({ success: false, msg: 'Trả xe thất bại. Vui lòng thử lại sau !' }))
-    })
-
-    app.get('/RentDetail/:IDDon/', async (req,res) => {
-        try{
-            const info = await SoXeModel.findOne({_id : `${req.params.IDDon}`}).populate("IDXe").populate("IDKH")
-            res.json(info)
-        }
-        catch(e){res.json(e)}
-       
-        // .catch(err => res.json(err))
-    })
-
+    app.use('/Rent', SoXeRoute)
 
     try{app.listen(port, () =>{console.log("Server khởi động tại port " + port)})}
     catch{console.log("Server khởi động thất bại")}
 }
-
-
-
