+++ src/pages/Dashboard.jsx
@@ -17,8 +17,8 @@ import CarMain from './CarManager/MainCar'
 import CarAdd from './CarManager/AddCar'
 import CarEdit from './CarManager/DetailCar'
 
-import Book from './BookManager/Book'
-import BookMain from './BookManager/BookMain'
+import Book from './BookManager/BookMain'
+import BookMain from './BookManager/BookChooseCar'
 import BookAdd from './BookManager/BookAdd'
 import BookDetail from './BookManager/BookDetail'
 import BookCreateRent from './BookManager/BookCreateRent'
